refactor packages structure

+++ src/main/java/com/spehof/piggy/service/MoneyHolderService.java
@@ -3,6 +3,7 @@ package com.spehof.piggy.service;
 import com.spehof.piggy.DAO.MoneyHolderDao;
 import com.spehof.piggy.domain.User;
 import com.spehof.piggy.domain.MoneyHolder;
+import com.spehof.piggy.exception.MoneyHolderStillUseException;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,10 +18,16 @@ import java.util.List;
 public class MoneyHolderService {
 
     private final MoneyHolderDao moneyHolderDao;
+    private final CostService costService;
+    private final EarningService earningService;
 
     @Autowired
-    public MoneyHolderService(MoneyHolderDao moneyHolderDao) {
+    public MoneyHolderService(MoneyHolderDao moneyHolderDao,
+                              CostService costService,
+                              EarningService earningService) {
         this.moneyHolderDao = moneyHolderDao;
+        this.costService = costService;
+        this.earningService = earningService;
     }
 
     public MoneyHolder create(User user, String title){
@@ -35,17 +42,27 @@ public class MoneyHolderService {
 
     public void delete(User user, MoneyHolder moneyHolderFromApi){
         MoneyHolder moneyHolderFromDb = user.getMoneyHolder(moneyHolderFromApi.getId());
+        if (this.checkUsing(moneyHolderFromDb)) throw new MoneyHolderStillUseException("Money holder " + moneyHolderFromDb.getTitle() + " still use");
         user.removeMoneyHolder(moneyHolderFromDb);
         moneyHolderDao.delete(moneyHolderFromDb);
     }
 
     public MoneyHolder update(User user, MoneyHolder moneyHolderFromApi){
         MoneyHolder moneyHolderFromDb = user.getMoneyHolder(moneyHolderFromApi.getId());
+        if (this.checkUsing(moneyHolderFromDb)) throw new MoneyHolderStillUseException("Money holder " + moneyHolderFromDb.getTitle() + " still use");
         if (moneyHolderFromApi.getTitle() == null){
-            BeanUtils.copyProperties(moneyHolderFromApi, moneyHolderFromDb, "id", "user");
+            BeanUtils.copyProperties(moneyHolderFromApi, moneyHolderFromDb, "id", "user", "title");
         } else {
             BeanUtils.copyProperties(moneyHolderFromApi, moneyHolderFromDb, "id", "user");
         }
         return moneyHolderDao.save(moneyHolderFromDb);
+    }
+
+    private boolean checkUsing(MoneyHolder moneyHolder){
+        if (costService.checkUsing(moneyHolder) || earningService.checkUsing(moneyHolder)) {
+            return true;
+        } else {
+            return false;
+        }
     }
 }
